Added project files and set up environment
Added LICENSE, .gitignore, pom.xml, and .editorconfig files. These are essential for the project setup. The LICENSE file clarifies the rules for the use of our software. The .gitignore allows us to specify files to be ignored by Git. The pom.xml is a configuration file for build automation. The .editorconfig helps maintain consistent coding styles.