Add Roman to Arabic conversion module and update POMs
Added a new module `roman-to-integer` to the `diffblue` project for converting Roman numerals to Arabic integers. This includes a `RomanToArabicConverter` class and corresponding POM file. Additionally, updated the parent POM files to include the new module. This change is to extend the functionality of the project to cover numeral systems conversion.