Add pom.xml for new module epam-ai-dial
Added a new pom.xml file for a fresh module called epam-ai-dial. This module is needed for Autonomous Java Unit Test Writing with AI for Code. In addition to the new file, the main pom.xml was edited to include new module and add XML and HTML report formats for Jacoco maven plugin.