Update pom.xml files with new details and dependencies
Updating `pom.xml` and `diffblue/pom.xml` with new developer and contributor information as well as modifying some dependencies. Jegors' email address has been updated, and a new role added. A new contributor, Alina, has been introduced. Furthermore, 'maven-site-plugin' was added to plugins. In `diffblue/pom.xml` file, a URL section was added and some dependencies were removed while junit-jupiter dependency was switched to junit-jupiter-api. This is done to keep up-to-date and precise information, and ensure dependencies align with the current needs of the project.