Reorganized project structure and renamed modules
Deleted files and directories related to the 'diffblue' and 'roman-to-integer' modules. These modules are part of a larger restructuring of the project. 'diffblue' and 'roman-to-integer' renamed to 'diffblue-cover' and 'diffblue-roman' respectively for better understanding of the role each module plays. The 'github-copilot' module created to introduce GitHub Copilot samples. Unit tests were also reorganised in accordance to the new hierarchy. Necessary changes in pom.xml were done to reflect the changes.

@coderabbitai ignore