Add legal age module with its predicate implementation and tests
This commit includes the creation of a new module 'diffblue-legal-age' in the main pom.xml file. A new class `LegalAgePredicate`, implementing the Predicate interface, was added to calculate if given birthdate pertains to legal age or not. A new 'pom.xml' file was added for the module and a test class 'LegalAgePredicateDiffblueTest' was added to include tests for the legal age predicate. This new functionality is essential for age verification in several parts of our application's logic.