Add RomanToArabicConverter and corresponding tests
Added RomanToArabicConverter class and its corresponding test class. The converter class is set to throw UnsupportedOperationException until it is implemented. The test class tests for the conversion of a wide range of roman numerals to their respective arabic numbers as well as checking that the converter responds appropriately to invalid roman numerals.