Remove pending test for invalid Roman numerals
This commit removes a pending test case from the RomanToArabicConverterTest.groovy file. The test case was initially designed to handle invalid Roman numerals. However, upon further consideration, it was concluded that this validation should be handled at a higher level, and not within the converter functionality. Hence, this 'result is undefined for invalid Roman numerals' test case is no longer needed in this module.